vault_scripts/requirements.txt
--index-url https://artifacts.dell.com/artifactory/api/pypi/python/simple
--trusted-host artifacts.dell.com


hvac
python-dotenv